split tests into two different files
